# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Motorway Petroleum API (MVP)
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /auth/login:
    post:
      summary: Login (returns JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: token
  /stations:
    get:
      summary: List stations
      responses:
        '200':
          description: array of stations
    post:
      summary: Create station
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
      responses:
        '201':
          description: station created
  /stations/{id}:
    get:
      summary: Get station
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: station
  /pumps:
    get:
      summary: List pumps
      responses:
        '200':
          description: array of pumps
  /transactions:
    post:
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                station_id:
                  type: string
                pump_id:
                  type: string
                liters:
                  type: number
                payment_method:
                  type: string
      responses:
        '201':
          description: transaction created